Move `DefaultSliderBody` to own class and make `SliderBody` more abstract
