+++ osu.Game.Rulesets.Osu/Skinning/SliderBody.cs
@@ -8,7 +8,6 @@ using System.Collections.Generic;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Lines;
-using osu.Game.Rulesets.Osu.Objects;
 using osu.Game.Rulesets.Osu.Skinning.Default;
 using osuTK;
 using osuTK.Graphics;
@@ -111,28 +110,6 @@ namespace osu.Game.Rulesets.Osu.Skinning
         /// <param name="vertices">The vertices</param>
         protected void SetVertices(IReadOnlyList<Vector2> vertices) => path.Vertices = vertices;
 
-        protected virtual DrawableSliderPath CreateSliderPath() => new DefaultDrawableSliderPath();
-
-        private partial class DefaultDrawableSliderPath : DrawableSliderPath
-        {
-            private const float opacity_at_centre = 0.3f;
-            private const float opacity_at_edge = 0.8f;
-
-            protected override Color4 ColourAt(float position)
-            {
-                const float actual_to_visual_radius = OsuHitObject.OBJECT_RADIUS / OsuHitObject.VISUAL_OBJECT_RADIUS;
-
-                position = 1 - (1 - position) * actual_to_visual_radius;
-
-                if (position < 0.0)
-                    return Color4.Transparent;
-
-                if (CalculatedBorderPortion != 0f && position <= CalculatedBorderPortion)
-                    return BorderColour;
-
-                position -= CalculatedBorderPortion;
-                return new Color4(AccentColour.R, AccentColour.G, AccentColour.B, (opacity_at_edge - (opacity_at_edge - opacity_at_centre) * position / GRADIENT_PORTION) * AccentColour.A);
-            }
-        }
+        protected abstract DrawableSliderPath CreateSliderPath();
     }
 }
